im-mesh-client-v2.1.0/requirements.txt
# Im Mesh Client Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Meshtastic
meshtastic>=2.5.0
pyserial>=3.5
Pypubsub>=4.0.3

# Database
aiosqlite==0.19.0

# Image Processing
Pillow>=10.0.0
heatshrink2>=0.12.0

# Utilities  
python-multipart==0.0.6
jinja2==3.1.2

# Data validation
pydantic>=2.5.0
